requirements.txt
amqtt
influxdb
django
djangorestframework
django-cors-headers